Refactored code to make it more modular and hava a slightly cleaner API
