+++ mmtf/api/__init__.py
@@ -1,0 +1,1 @@
+from mmtf.api.default import parse,parse_gzip,fetch,MMTFDecoder
