+++ mmtf/codecs/codecs.py
