+++ mmtf/encoders/__init__.py
@@ -1,0 +1,1 @@
+from mmtf.encoders.encoders import run_length_encode,delta_encode
